[docs][fix]: 补充 HAL bind 一致性说明，修复 drv_vfd 符号比较编译错误
README 边界规则里补充三个组合层 bind 独立函数的一致性例证；
VFD_GEAR_MAX 改为无符号字面量后 drv_vfd_apply_gear 的比较需同步加类型转换。

+++ framework/adapters/outbound/hal/providers/snack/modbus/drv_vfd.c
@@ -235,7 +235,7 @@ sw_err_t drv_vfd_apply_gear(drv_vfd_t *vfd, hal_vfd_gear_t gear)
 
     abs_gear_int = vfd_abs_gear(gear);
 
-    if (abs_gear_int > VFD_GEAR_MAX) {
+    if (abs_gear_int > (int)VFD_GEAR_MAX) {
         return SW_ERR_PARAM;
     }
     if ((gear < 0) && !vfd_has_rev(vfd)) {
